Added dedicated file for constants
