+++ src/index.ts
@@ -579,4 +579,5 @@ const CLEAN_ROOT_MODEL = omit(ROOT_MODEL, ['system']) as PublicModel;
 export { Transaction, CLEAN_ROOT_MODEL as ROOT_MODEL };
 
 // Expose the main error class and query symbols
-export { RoninError, QUERY_SYMBOLS, getQuerySymbol } from '@/src/utils/helpers';
+export { RoninError, getQuerySymbol } from '@/src/utils/helpers';
+export { QUERY_SYMBOLS } from '@/src/utils/constants';
