+++ tests/instructions/including.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'bun:test';
-import { type Model, type Query, Transaction } from '@/src/index';
+import { type Model, QUERY_SYMBOLS, type Query, Transaction } from '@/src/index';
 
 import {
   RECORD_ID_REGEX,
@@ -7,7 +7,6 @@ import {
   queryEphemeralDatabase,
 } from '@/fixtures/utils';
 import type { MultipleRecordResult, SingleRecordResult } from '@/src/types/result';
-import { QUERY_SYMBOLS } from '@/src/utils/helpers';
 
 test('get single record including unrelated record without filter', async () => {
   const queries: Array<Query> = [
