+++ tests/instructions/to.test.ts
@@ -1,9 +1,15 @@
 import { expect, test } from 'bun:test';
-import { type Model, type Query, type StoredObject, Transaction } from '@/src/index';
+import {
+  type Model,
+  QUERY_SYMBOLS,
+  type Query,
+  RoninError,
+  type StoredObject,
+  Transaction,
+} from '@/src/index';
 
 import { RECORD_TIMESTAMP_REGEX, queryEphemeralDatabase } from '@/fixtures/utils';
 import type { MultipleRecordResult, SingleRecordResult } from '@/src/types/result';
-import { QUERY_SYMBOLS, RoninError } from '@/src/utils/helpers';
 
 test('set single record to new string field', async () => {
   const queries: Array<Query> = [
