+++ tests/instructions/using.test.ts
@@ -1,5 +1,11 @@
 import { expect, test } from 'bun:test';
-import { type Model, type Query, Transaction } from '@/src/index';
+import {
+  type Model,
+  QUERY_SYMBOLS,
+  type Query,
+  RoninError,
+  Transaction,
+} from '@/src/index';
 
 import {
   RECORD_ID_REGEX,
@@ -7,7 +13,6 @@ import {
   queryEphemeralDatabase,
 } from '@/fixtures/utils';
 import type { SingleRecordResult } from '@/src/types/result';
-import { QUERY_SYMBOLS, RoninError } from '@/src/utils/helpers';
 
 test('get single record using preset', async () => {
   const queries: Array<Query> = [
